// lib/type.ts
import {z} from 'zod';

export const FormSchema = z.object({
  name: z.string().min(1, 'Name is required'),
  email: z.string().email('Invalid email address'),
  subject: z.string().min(1, 'Subject is required'),
  message: z.string().min(1, 'Message is required'),
});
export type ActionResult = { error: string } | { success: true };